==> 1_3_predict_gender.ipynb <==
Use the library from here: https://github.com/alt-code/Research/tree/master/SimpleGenderComputer

Please use it with Python 2.7

import pickle
import pandas as pd
from simpleGenderComputer import SimpleGenderComputer

# get the data from the library provided in the first link.

gc = SimpleGenderComputer('./namelists')

def get_author_name(name):
    words = name.split()
    # must has first and last name to query the API
    if len(words) <= 1:
        payload = 'invalid'
    else:
        payload = {'Fname': words[0], 'Lname': words[-1]}
    return payload

### Authors in self-promotion data

! wc -l ./author_names.txt

name_gender = {}
with open('./author_names.txt', 'r') as ifile:
    for name in ifile:
        # name is already normalized
        # get rid of '\n' at the end of each line
        name = name.strip()
        name_tupe = get_author_name(name)
        if type(name_tupe) == dict:
            fname = name_tupe['Fname']
            fname = fname.decode('utf-8', 'ignore')
            gender = gc.simpleLookup(fname)
            name_gender[name] = gender

len(name_gender)

with open("./name_gender.txt", 'w') as ofile:
    for name, gender in name_gender.items():
        ofile.write(name + '||' + gender + '\n')

### Authors in IOP data

iop_clean_name_self_gender = pd.read_csv('./iop_clean_name_self_gender.csv')

def pred_gender_ford(name):
    if type(name) == str:
        name = name.strip()
        name_tupe = get_author_name(name)
        if type(name_tupe) == dict:
            fname = name_tupe['Fname']
            fname = fname.decode('utf-8', 'ignore')
            if len(fname) > 1:
                gender = gc.simpleLookup(fname)
                return gender
            else:
                return 'invalid'
        else:
            return 'invalid'
    else:
        return 'invalid'

iop_clean_name_self_gender['gender_ford'] = iop_clean_name_self_gender['clean_name'].apply(pred_gender_ford)

iop_clean_name_self_gender.to_csv("./iop_clean_name_self_gender_ford.csv", index=False, header=True, encoding='utf-8')
