==> 2_reg_data_mentions_audiences_eth.ipynb <==
import os
import math
import json
import csv
import time
import random
import requests
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.colors
import matplotlib.cm as cm
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
from datetime import datetime

Data_Root = '/Data/Promotion/'

# pd.options.display.precision = 20

CIs = {'90': 1.645, '95': 1.96, '99': 2.576}

labels = ['Male', 'Female']

colors = sns.color_palette()[:len(labels)]

# load reg data with basic controls
utype = {'author_id': str, 'matched_tid': 'str', 'matched_tid_original': 'str', 'matched_tid_retweet': 'str'}
reg_data = pd.read_csv(Data_Root+"revision/reg_data.csv", header=0, dtype=utype)

# before dropping missing values
reg_data.shape

reg_data.isna().sum()

# num of obs in regression after dropping missing data
inx = reg_data[reg_data.columns[~reg_data.columns.isin(['matched_tid', 'matched_tid_original', 'matched_tid_retweet'])]].dropna().index


# mainly dropped due to missing data on affiliation and journal impact.
len(inx)

reg_data = reg_data.loc[inx]
reg_data.index = range(len(reg_data))

del inx

### Add total mentions and self-promotion position

dois = set(reg_data.doi)

len(dois)

doi_tids = {}
with open(Data_Root+'revision/papers_2013_2018.json', 'r') as ifile:
    for line in ifile:
        res = json.loads(line)
        doi = res['doi']
        if doi in dois:
            doi_tids[doi] = [tweet['tweet_id'] for tweet in res['data']['tweets']]

len(doi_tids)

tids = set()
for doi in doi_tids:
    for tid in doi_tids[doi]:
        tids.add(tid)

len(tids)

tid_date_all = {}
tid_follow_cn_all = {}

with open(Data_Root+'tweets_v2.json', 'r') as ifile:
    for line in ifile:
        line = json.loads(line)
        tid, tdate = line['id_str'], line['created_at']
        if tid in tids:
            tid_date_all[tid] = datetime.strptime(tdate, '%a %b %d %H:%M:%S +0000 %Y')#.strftime("%Y-%m-%d")
            tid_follow_cn_all[tid] = line['user']['followers_count']

# this file is collected earlier, but without full text
with open(Data_Root+'tweets_v2_non_full_text.json', 'r') as ifile:
    for line in ifile:
        line = json.loads(line)
        tid, tdate = line['id_str'], line['created_at']
        if tid in tids:
            # tid_name_all[tid] = line['user']['name']
            tid_date_all[tid] = datetime.strptime(tdate, '%a %b %d %H:%M:%S +0000 %Y')#.strftime("%Y-%m-%d")
            tid_follow_cn_all[tid] = line['user']['followers_count']

len(tid_date_all)

len(tid_follow_cn_all)

Count num of tweets for the final set of papers

# focus on Male and Female
len(reg_data.loc[reg_data['gender'].isin(['Female', 'Male'])])

cn = 0
for doi in set(reg_data.loc[reg_data['gender'].isin(['Female', 'Male'])]['doi']):
    cn += len(doi_tids[doi])

cn

def check_early_self_promo(row):
    doi, p_0, mat_tid = row['doi'], row['self_promotion'], row['matched_tid']
    valid_tids = [tid for tid in doi_tids[doi] if tid in tid_date_all]
    total = len(valid_tids)
    if p_0:
        sort_tids = sorted(valid_tids, key=lambda tid: tid_date_all[tid])
        # valid_users = [tid_name_all[tid] for tid in sort_tids]
        # ix = valid_users.index(t_user)
        ix = sort_tids.index(mat_tid) + 1
        return (total, ix, ix/total)
    else:
        return (total, -1, -1)

reg_data['Tem'] = reg_data.apply(lambda row: check_early_self_promo(row), axis = 1)
reg_data[['total_num_tweets', 'self_promo_position', 'self_promo_position_frac']] = pd.DataFrame(reg_data.Tem.values.tolist(), index = reg_data.index)
reg_data = reg_data.drop(columns=['Tem'])


np.sum(reg_data['self_promo_position'] == 1)

np.sum(reg_data['self_promo_position'] == 1) / np.sum(reg_data['self_promotion'] == True)

### Add num. of followers control

def get_tid_follower_cn(row):
    doi, p_0, mat_tid = row['doi'], row['self_promotion'], row['matched_tid']
    if p_0:
        return tid_follow_cn_all[mat_tid]
    else:
        return -1

reg_data['matched_tid_follower_cn'] = reg_data.apply(get_tid_follower_cn, axis=1)

reg_data.head()

### Add num of retweets for self-promotional tweets

len(reg_data)

len(reg_data.loc[reg_data['self_promotion'] == True])

np.sum(reg_data.loc[reg_data['self_promotion'] == True]['total_num_tweets'].isna())

len(reg_data.loc[~reg_data['matched_tid'].isna()])

tids_ = set(reg_data.loc[~reg_data['matched_tid'].isna()]['matched_tid'])

tid_engage_stats = {}

with open(Data_Root+'tweets_v2.json', 'r') as ifile:
    for line in ifile:
        line = json.loads(line)
        tid = line['id_str']
        if tid in tids_:
            tid_engage_stats[tid] = (line['retweet_count'], line['favorite_count'])

# this file is collected earlier
with open(Data_Root+'tweets_v2_non_full_text.json', 'r') as ifile:
    for line in ifile:
        line = json.loads(line)
        tid = line['id_str']
        if tid in tids_:
            tid_engage_stats[tid] = (line['retweet_count'], line['favorite_count'])

reg_data['matched_tid_retweet_cn'] = reg_data['matched_tid'].apply(lambda tid: tid_engage_stats[tid][0] if tid in tid_engage_stats else 0)
reg_data['matched_tid_likes_cn'] = reg_data['matched_tid'].apply(lambda tid: tid_engage_stats[tid][1] if tid in tid_engage_stats else 0)

reg_data.head()

len(reg_data.loc[reg_data['matched_tid_retweet_cn'] != -1])

### Different definition of self-promotion

Define self-promotion timing as the actual time difference between promotion time and publication date

doi_pub_date = {}
doi_tw_aud = {}

cn_lines = 0
with open(Data_Root + 'merged.txt', 'r') as ofile:
    for row in ofile:
        cn_lines += 1
        if cn_lines % 1000000 == 0:
            print('processed %d lines...'%cn_lines)
        if '}{"altmetric_id"' in row:
            row = row.replace('}{"altmetric_id"', '}\n{"altmetric_id"')
        # this way works for both good and bad lines.
        records = row.split('\n')
        for record in records:
            if record != '':
                paper = json.loads(record)
                citation = paper['citation']
                pubdate = ''
                if 'pubdate' in citation:
                    pubdate = citation['pubdate']
                elif 'epubdate' in citation:
                    pubdate = citation['epubdate']
                doi = ''
                if 'doi' in citation:
                    doi = citation['doi']
                if doi in dois:
                    doi_pub_date[doi] = pubdate
                    if 'poster_types' in paper['demographics']:
                        doi_tw_aud[doi] = paper['demographics']['poster_types']

def cal_self_promo_time(row):
    doi, p_0, mat_tid = row['doi'], row['self_promotion'], row['matched_tid']
    if p_0:
        promo_time = tid_date_all[mat_tid]
        pub_time = doi_pub_date[doi]
        pub_time = datetime.strptime(pub_time, '%Y-%m-%dT%H:%M:%S+00:00')
        delta = (promo_time - pub_time).total_seconds()
        # promo time often happens before recorded pub date.
        if delta <= 0:
            return 0
        else:
            return delta
    else:
        return -1

reg_data['self_promo_time'] = reg_data.apply(lambda row: cal_self_promo_time(row), axis = 1)

reg_data.loc[reg_data['self_promo_time'] != -1]

np.median(reg_data[reg_data.self_promotion == True]['self_promo_time'])

np.mean(reg_data[reg_data.self_promotion == True]['self_promo_time'])

np.mean(reg_data.total_num_tweets)

Consider promotion within one day after publication as self-promotion

day_sec = 3600*24
thsh = 1

def promo_fun(delta):
    if delta == -1:
        return False
    else:
        if delta <= day_sec * thsh:
            return True
        else:
            return False

reg_data['self_promotion_def'] = reg_data['self_promo_time'].apply(promo_fun)

reg_data.loc[reg_data['self_promotion_def'] == True, ['self_promo_time', 'self_promotion_def']].sample(100)

Code DV as whether the author is the first one to self-promote for a paper

doi_first_aid = {}

for doi, gp in reg_data.groupby('doi'):
    gp = gp.loc[gp['self_promo_time'] != -1]
    if len(gp) >= 1:
        aid, etime = sorted(gp[['author_id', 'self_promo_time']].values.tolist(), key=lambda x: x[1])[0]
        doi_first_aid[doi] = aid

len(doi_first_aid)

len(set(reg_data[reg_data['self_promotion'] == True]['doi']))

def is_first_to_promo(row):
    doi, aid = row['doi'], row['author_id']
    if doi in doi_first_aid:
        if aid == doi_first_aid[doi]:
            return True
        else:
            return False
    else:
        return False

reg_data['self_promotion_first'] = reg_data.apply(lambda row: is_first_to_promo(row), axis = 1)

sum(reg_data['self_promotion_first']==True)

### Add audiences (DV)

len(dois)

len(doi_tw_aud)

doi_tw_aud['10.1001/jama.2018.17121']

def get_num_audiences(doi, kind = 'member_of_the_public'):
    if doi in doi_tw_aud and kind in doi_tw_aud[doi]:
        return doi_tw_aud[doi][kind]
    else:
        return 0

reg_data['member_of_the_public'] = reg_data['doi'].apply(lambda doi: get_num_audiences(doi, 'member_of_the_public'))
reg_data['researcher'] = reg_data['doi'].apply(lambda doi: get_num_audiences(doi, 'researcher'))
reg_data['practitioner'] = reg_data['doi'].apply(lambda doi: get_num_audiences(doi, 'practitioner'))
reg_data['science_communicator'] = reg_data['doi'].apply(lambda doi: get_num_audiences(doi, 'science_communicator'))

reg_data['num_non_scientists'] = reg_data[['member_of_the_public', 'practitioner', 'science_communicator']].sum(axis=1)


np.sum(reg_data['researcher'])

np.sum(reg_data['num_non_scientists'])

### Add author ethnicity (focus on male and female to reduce the number of authors queried)

eth_cat_map = {
 'ISRAELI': 'MiddleEastern',
 'ARAB': 'MiddleEastern',
 'TURKISH': 'MiddleEastern',
 'ENGLISH': 'English',
 'HISPANIC': 'SouthernEuropean',
 'ITALIAN': 'SouthernEuropean',    
 'GREEK': 'SouthernEuropean',
 'GERMAN': 'WesternNorthernEuropean',
 'NORDIC': 'WesternNorthernEuropean',
 'DUTCH': 'WesternNorthernEuropean',
 'FRENCH': 'WesternNorthernEuropean',
 'BALTIC': 'WesternNorthernEuropean',
 'HUNGARIAN': 'EasternEuropean',
 'ROMANIAN': 'EasternEuropean',
 'SLAV': 'EasternEuropean',
 'CHINESE': 'Chinese',
 'INDIAN': 'Indian',
 'AFRICAN': 'African',
 'KOREAN': 'EastAsian',
 'JAPANESE': 'EastAsian',
 'THAI': 'EastAsian',
 'VIETNAMESE': 'EastAsian',
 'INDONESIAN': 'EastAsian',
 'MONGOLIAN': 'EastAsian',
 'CARIBBEAN': 'CARIBBEAN',
 'POLYNESIAN': 'POLYNESIAN',
 'org': 'org',
 'unknown': 'unknown'}

def parse_author_name(name):
    words = name.split()
    # must has first and last name to query the API
    if len(words) <= 1:
        payload = 'invalid'
    else:
        payload = {'Fname': words[0], 'Lname': words[-1]}
        # C. Kirabo Jackson
        if len(words) >= 3:
            # C or C.
            if len(words[0]) == 1 or (len(words[0]) == 2 and words[0][1] == '.'):
                # not C. J. Jackson; but C. Del Jackson
                if len(words[1]) > 2 or (len(words[1]) == 2 and words[1][1] != '.'):
                    given = words[1]
                    family = words[-1]
                    payload = {'Fname': given, 'Lname': family}
    return payload

name_filter = set(['consortium', 'collaboration', 'collaboration*', 'editor', 'bank'])

def get_author_eth_gen(name):
    # invalid: single-word-name
    feats = ['unknown', 'unknown']
    name_tupe = parse_author_name(name)
    if type(name_tupe) == dict:
        family = name_tupe['Lname']
        # The XXX Collaboration would go to API as well since the Lname is Collaboration.
        if family.lower() in name_filter:
            feats = ['org', 'org']
        else:
            if name in name_eth_gen:
                # ethnicity
                major = name_eth_gen[name]['Ethnea'].split('-')[0]
                if major != "ERROR" and major != 'UNKNOWN' and major != 'TOOSHORT':
                    feats[0] = major
                # gender
                gender = name_eth_gen[name]['Genni']
                if gender != '-':
                    feats[1] = gender
    return feats

# focus on Male and Female
len(reg_data.loc[reg_data['gender'].isin(['Female', 'Male'])])

len(set(reg_data.loc[reg_data['gender'].isin(['Female', 'Male'])]['author_name']))

eth_url = 'http://abel.lis.illinois.edu/cgi-bin/ethnea/search.py?format=json'

name_eth_gen = {}

# reuse data collected in the initial submission
with open(Data_Root+'name_eth_gen.json', 'r') as ofile:
    for row in ofile:
        row = json.loads(row)
        name, info = row['name'], row['info']
        name_eth_gen[name] = info

len(name_eth_gen)

# focus on Male and Female
ix = 0
for name in set(reg_data.loc[reg_data['gender'].isin(['Female', 'Male'])]['author_name']):
    if name not in name_eth_gen:
        payload = parse_author_name(name)
        if payload != 'invalid':
            ix += 1
            if ix%5000 == 0:
                time.sleep(5)
            response = requests.get(eth_url, params=payload)
            try:
                j = eval(response.text)
                genni = j['Genni']
                ethnea = j['Ethnea']
                name_eth_gen[name] = {'Ethnea': ethnea, 'Genni': genni}
            except:
                continue
            if ix%100000 == 0:
                print(ix)

with open(Data_Root+'revision/name_eth_gen.json', 'w') as ofile:
    for name, info in name_eth_gen.items():
        row = {'name': name, 'info': info}
        ofile.write(json.dumps(row) + '\n')

name_eth_gen = {}

with open(Data_Root+'revision/name_eth_gen.json', 'r') as ofile:
    for row in ofile:
        row = json.loads(row)
        name, info = row['name'], row['info']
        name_eth_gen[name] = info

len(name_eth_gen)

Add inferred ethnicity (based on Ethnea)

reg_data['author_eth_gen'] = reg_data['author_name'].apply(lambda name: get_author_eth_gen(name))
reg_data[['author_eth_ethnea', 'author_gender_ethnea']] = pd.DataFrame(reg_data['author_eth_gen'].values.tolist(), index = reg_data.index)
reg_data = reg_data.drop(columns = ['author_eth_gen'])
reg_data['author_eth_ethnea_broad'] = reg_data['author_eth_ethnea'].map(eth_cat_map)


len(reg_data)

reg_data.head(3)

reg_data.columns

reg_data.to_csv(Data_Root+"revision/reg_data_drop_missing.csv", index=False, header=True, encoding='utf-8')

len(reg_data)

# utype = {'author_id': str, 'matched_tid': str, 'matched_tid_original': str, 'matched_tid_retweet': str}
# reg_data = pd.read_csv(Data_Root + "revision/reg_data_drop_missing.csv", header=0, dtype=utype)
