==> 4_reg_data_with_tw_control.ipynb <==
import os
import math
import json
import csv
import time
import random
import requests
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.colors
import matplotlib.cm as cm
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
from collections import defaultdict, Counter
from datetime import datetime

Data_Root = '/Data/Promotion/'

# pd.options.display.precision = 20

CIs = {'90': 1.645, '95': 1.96, '99': 2.576}

labels = ['Male', 'Female']

colors = sns.color_palette()[:len(labels)]

utype = {'author_id': str, 'matched_tid': str, 'matched_tid_original': str, 'matched_tid_retweet': str}
reg_data = pd.read_csv(Data_Root + "revision/reg_data_drop_missing.csv", header=0, dtype=utype)

reg_data.shape

reg_data.head(3)

Add paper publication date

doi_pub_date = {}
with open(Data_Root+'revision/papers_2013_2018.json', 'r') as ifile:
    for line in ifile:
        res = json.loads(line)
        doi_pub_date[res['doi']] = res['data']['pub_date']

len(doi_pub_date)

doi_pub_date['10.4202/app.00261.2016']

reg_data['pub_date'] = reg_data['doi'].apply(lambda doi: doi_pub_date[doi][:10])

# https://docs.openalex.org/about-the-data
# manually check these ids on openAlex: https://api.openalex.org/authors/A2569941301
reg_data[['author_id', 'author_name']]

# reg_data.isna().sum()

### Using author-tweeter matching data from the external source

Data: `Costas et al. 2020 - Large-scale identification and characterization of scholars on Twitter`

aid_uid_df = pd.read_csv(Data_Root + "revision/authors_tweeters_2022_08_21.csv", header=0, \
                        dtype={'tweeter_id': str})

len(aid_uid_df)

aid_uid_df['mag_aid'] = aid_uid_df['author_id'].apply(lambda url: url.split('https://openalex.org/A')[1])

aid_uid_df.sample(10)

aid_uid_df.isna().sum()

aid_uid_df['valid'].value_counts(dropna=False)

np.mean(aid_uid_df['valid'])

len(aid_uid_df.loc[aid_uid_df['valid'] != 0])

aid_uids = {}

for aid, gp in aid_uid_df.loc[aid_uid_df['valid'] != 0].groupby('mag_aid'):
    aid_uids[aid] = list(set(gp['tweeter_id']))

len(aid_uids)

len(set(reg_data.loc[reg_data['author_id'].isin(aid_uids)]['author_id']))

cn = 0
for aid in aid_uids:
    if len(aid_uids[aid]) > 1:
        cn += 1

# some authors have matched to multiple tw accounts.
cn

all_uids = set()

for aid in aid_uids:
    for uid in aid_uids[aid]:
        all_uids.add(uid)

# num of unique Tw user id
len(all_uids)

uid_account_creation_time_follower = {}
uid_active_times = defaultdict(list)

with open(Data_Root+'tweets_v2.json', 'r') as ifile:
    for line in ifile:
        line = json.loads(line)
        uid = line['user']['id_str']
        if uid in all_uids:
            if uid not in uid_account_creation_time_follower:
                creation_time = line['user']['created_at']
                # this is counted at data collection time. All tweets are collected around the same time.
                f_cn = line['user']['followers_count']
                uid_account_creation_time_follower[uid] = (creation_time, f_cn)
            tweet_time = line['created_at']
            uid_active_times[uid].append(tweet_time)

with open(Data_Root+'tweets_v2_non_full_text.json', 'r') as ifile:
    for line in ifile:
        line = json.loads(line)
        uid = line['user']['id_str']
        if uid in all_uids:
            if uid not in uid_account_creation_time_follower:
                creation_time = line['user']['created_at']
                f_cn = line['user']['followers_count']
                uid_account_creation_time_follower[uid] = (creation_time, f_cn)
            tweet_time = line['created_at']
            uid_active_times[uid].append(tweet_time)

uid_active_times = dict(uid_active_times)

count how many of their uids are in our tw data.

len(uid_account_creation_time_follower)

len(uid_active_times)

with open(Data_Root+'revision/uid_account_info.json', 'w') as ofile:
    for uid in uid_active_times:
        creation_time, f_cn = uid_account_creation_time_follower[uid]
        alist = uid_active_times[uid]
        res = {'tweeter_id': uid, 'created_at': creation_time, 'follower_cn_snapshot': f_cn, 'active_time': alist}
        ofile.write(json.dumps(res)+'\n')

for uid, alist in uid_active_times.items():
    blist = []
    for tdate in alist:
        tdate = datetime.strptime(tdate, '%a %b %d %H:%M:%S +0000 %Y')
        # reformat back to string
        # tdate = tdate.strftime("%Y-%m-%dT%H:%M:%S")
        blist.append(tdate)
    uid_active_times[uid] = sorted(blist)

uid_active_times['11100372']

def code_tw_active_status_fcn(row):
    aid, pub_date = row['author_id'], row['pub_date']
    pub_date = datetime.strptime(pub_date, '%Y-%m-%d')
    flag = False
    fcn = -1
    if aid in aid_uids:
        for uid in aid_uids[aid]:
            # get the earliest active date
            if uid in uid_active_times:
                eatime = uid_active_times[uid][0]
                if (pub_date - eatime).total_seconds() >= 0:
                    flag = True
                    fcn = uid_account_creation_time_follower[uid][1]
                    break
    return (flag, fcn)

# an author_id could be active for some of their observations, but inactive for other obs.
reg_data['Tem'] = reg_data.apply(lambda row: code_tw_active_status_fcn(row), axis = 1)

reg_data[['is_active_on_twitter', 'follower_cn_snapshot']] = pd.DataFrame(reg_data['Tem'].values.tolist(), index = reg_data.index)
reg_data = reg_data.drop(columns=['Tem'])


reg_data['is_active_on_twitter'].value_counts(dropna=False)

# self_promotion is coded based on our heuristic, active status is coded based on the external data.
# discrepency is expected due to different heuristics.
reg_data.loc[reg_data['self_promotion'] == True]['is_active_on_twitter'].value_counts()

np.median(reg_data.loc[reg_data['is_active_on_twitter']==True, 'follower_cn_snapshot'])

for year, gp in reg_data.loc[reg_data['is_active_on_twitter'] == True].groupby('pub_year'):
    print(year, np.mean(gp['self_promotion']))
