==> 5_psm_analysis.ipynb <==
import os
import math
import json
import numpy as np
import pandas as pd
import seaborn as sns
from collections import Counter
import matplotlib.colors
import matplotlib.cm as cm
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
from datetime import datetime
# from patsy import dmatrices
from patsy import dmatrix
from scipy.stats import ttest_rel
from scipy.stats import ttest_ind
import statsmodels.formula.api as smf

from causalinference import CausalModel
from causalinference.utils import random_data

Data_Root = '/Data/Promotion/'

CIs = {'90': 1.645, '95': 1.96, '99': 2.576}

labels = ['Male', 'Female']

colors = sns.color_palette()[:len(labels)]

reg_data = pd.read_csv(Data_Root+"revision/reg_data_drop_missing.csv", header=0, dtype={'matched_tid': str, \
        'matched_tid_retweet': str, 'matched_tid_original': str, 'pub_year': str})

reg_data = reg_data.loc[reg_data['gender'].isin(['Female', 'Male'])]
reg_data = reg_data.loc[reg_data.affiliation_cate != 'unknown']
reg_data.index = range(len(reg_data))

len(reg_data)

reg_data['author_citation_log'] = reg_data['author_citation'].apply(lambda x: np.log2(x+1))

reg_data['gender_CT'] = reg_data['gender'].apply(lambda gen: 1 if gen == 'Female' else 0)

co_feats_ = ["C(authorship_pos, Treatment(reference='last_position'))", 'author_pub_count_cate', 'affiliation_rank_cate', 'affiliation_cate', 'num_authors', \
         'journal_impact', 'author_citation_log', 'pub_year', \
        'Social_Sciences', 'Materials_Science', 'Engineering', 'Chemistry',
       'Biochemistry__Genetics_and_Molecular_Biology', 'Medicine', 'Nursing',
       'Agricultural_and_Biological_Sciences',
       'Pharmacology__Toxicology_and_Pharmaceutics', 'Neuroscience',
       'Business__Management_and_Accounting',
       'Economics__Econometrics_and_Finance', 'Chemical_Engineering',
       'Physics_and_Astronomy', 'Computer_Science', 'Decision_Sciences',
       'Health_Professions', 'Psychology', 'Immunology_and_Microbiology',
       'Dentistry', 'Earth_and_Planetary_Sciences', 'Environmental_Science',
       'Mathematics', 'Arts_and_Humanities', 'Energy', 'Veterinary', 'General']

X = dmatrix(formula_like=' + '.join(co_feats_), data=reg_data, return_type="dataframe")

list(X.columns)

X['gender'] = reg_data['gender']
X['gender_CT'] = reg_data['gender_CT']
X['self_promotion'] = reg_data['self_promotion']

co_feats = [
 "C(authorship_pos, Treatment(reference='last_position'))[T.first_position]",
 "C(authorship_pos, Treatment(reference='last_position'))[T.middle_position]",
 "C(authorship_pos, Treatment(reference='last_position'))[T.solo_author]",
 'affiliation_cate[T.international]',
 'pub_year[T.2014]',
 'pub_year[T.2015]',
 'pub_year[T.2016]',
 'pub_year[T.2017]',
 'pub_year[T.2018]',
 'author_pub_count_cate',
 'affiliation_rank_cate',
 'num_authors',
 'journal_impact',
 'author_citation_log',
 'Social_Sciences',
 'Materials_Science',
 'Engineering',
 'Chemistry',
 'Biochemistry__Genetics_and_Molecular_Biology',
 'Medicine',
 'Nursing',
 'Agricultural_and_Biological_Sciences',
 'Pharmacology__Toxicology_and_Pharmaceutics',
 'Neuroscience',
 'Business__Management_and_Accounting',
 'Economics__Econometrics_and_Finance',
 'Chemical_Engineering',
 'Physics_and_Astronomy',
 'Computer_Science',
 'Decision_Sciences',
 'Health_Professions',
 'Psychology',
 'Immunology_and_Microbiology',
 'Dentistry',
 'Earth_and_Planetary_Sciences',
 'Environmental_Science',
 'Mathematics',
 'Arts_and_Humanities',
 'Energy',
 'Veterinary',
 'General']

# Y is the outcome, D is treatment status, and X is the independent variable
causal = CausalModel(Y=X['self_promotion'].values, D=X['gender_CT'].values, \
                     X=X[co_feats].values)

print(causal.summary_stats)

causal.est_propensity()

# Automated propensity score estimation of which features to use in the matching.
# causal.est_propensity_s()

# Propensity model results
print(causal.propensity)

causal.propensity['fitted']

### 1 on 1 matching using the nearest neighbor

Check covariate balance before matching

for feat in co_feats:
    print(feat)
    f_avg = np.mean(X.loc[X['gender']=='Female', feat])
    m_avg = np.mean(X.loc[X['gender']=='Male', feat])
    print('\tFemale:\t', f_avg)
    print('\tMale:\t', m_avg)
    print('\tDiff:\t', f_avg-m_avg)
    print('\tT-test:\t', ttest_ind(X.loc[X['gender']=='Female', feat], X.loc[X['gender']=='Male', feat])[1])

len(causal.propensity['fitted'])

X['pscore'] = causal.propensity['fitted']

tem = X[['gender', 'pscore']].sort_values(by = ['pscore'])
tem['index'] = tem.index

tem = tem.values.tolist()

tem[-10:]

leng = len(tem)

pairs = {}

for i, elms in enumerate(tem):
    gen, score, ix = elms
    if gen == 'Female':
        pre_male_ix, nex_male_ix = 0, 0
        j = i-1
        while j >= 0:
            if tem[j][0] != 'Male':
                j -= 1
            else:
                break
        if j >= 0:
            pre_male_ix = j
        n = i+1
        while n <= leng-1:
            if tem[n][0] != 'Male':
                n += 1
            else:
                break
        if n <= leng-1:
            nex_male_ix = n
        if abs(score - tem[pre_male_ix][1]) <= abs(score - tem[nex_male_ix][1]):
            pairs[ix] = tem[pre_male_ix][2]
        else:
            pairs[ix] = tem[nex_male_ix][2]

del tem

len(pairs)

Check covariate balance after matching

for feat in co_feats:
    print('\n')
    print('Feat:', feat, '\n')
    print('Before matching:\n')
    f_avg = np.mean(X.loc[X['gender']=='Female', feat])
    m_avg = np.mean(X.loc[X['gender']=='Male', feat])
    print('\tFemale:\t', f_avg)
    print('\tMale:\t', m_avg)
    print('\tDiff:\t', f_avg-m_avg)
    print('\tT-test:\t', ttest_ind(X.loc[X['gender']=='Female', feat], X.loc[X['gender']=='Male', feat])[1])

    print('\nAfter matching:\n')
    f_avg = np.mean(X.loc[pairs.keys(), feat])
    m_avg = np.mean(X.loc[pairs.values(), feat])
    print('\tFemale:\t', f_avg)
    print('\tMale:\t', m_avg)
    print('\tDiff:\t', f_avg-m_avg)
    print('\tT-test:\t', ttest_rel(X.loc[pairs.keys(), feat], X.loc[pairs.values(), feat])[1])

for feat in co_feats:
    f_avg = np.mean(X.loc[X['gender']=='Female', feat])
    m_avg = np.mean(X.loc[X['gender']=='Male', feat])
    f_avg_ = np.mean(X.loc[pairs.keys(), feat])
    m_avg_ = np.mean(X.loc[pairs.values(), feat])
    print(feat, ' & ', '{:6.2f}'.format(f_avg), ' & ', '{:6.2f}'.format(m_avg), ' & ', '{:6.2f}'.format(f_avg_), ' & ', '{:6.2f}'.format(m_avg_), ' \\\\ \hline')

Avg treatment effect

# female avg
np.mean(X.loc[pairs.keys(), 'self_promotion'])

# male avg
np.mean(X.loc[pairs.values(), 'self_promotion'])

ttest_rel(X.loc[pairs.keys(), 'self_promotion'].apply(lambda x: 1 if x == True else 0), \
          X.loc[pairs.values(), 'self_promotion'].apply(lambda x: 1 if x == True else 0))

Avg treatment effect: `estimate via logit with covariates to control for the slight remaining imbalance`

# https://stats.stackexchange.com/questions/406379/how-exactly-to-evaluate-treatment-effect-after-matching

reg_data['self_promotion'] = reg_data['self_promotion'].astype('int32')

# eq = 'self_promotion ~ gender + ' + ' + '.join(co_feats_)

eq = "self_promotion ~ C(gender, Treatment(reference='Male')) + C(authorship_pos, Treatment(reference='last_position')) + author_pub_count_cate + \
    I(author_pub_count_cate**2) + affiliation_rank_cate + affiliation_cate + num_authors + journal_impact + author_citation_log + pub_year + \
        Social_Sciences + Materials_Science + Engineering + Chemistry + \
        Biochemistry__Genetics_and_Molecular_Biology + Medicine + Nursing + Agricultural_and_Biological_Sciences + \
        Pharmacology__Toxicology_and_Pharmaceutics + Neuroscience + Business__Management_and_Accounting + \
        Economics__Econometrics_and_Finance + Chemical_Engineering + Physics_and_Astronomy + Computer_Science + \
        Decision_Sciences + Health_Professions + Psychology + Immunology_and_Microbiology + Dentistry + \
        Earth_and_Planetary_Sciences + Environmental_Science + Mathematics + Arts_and_Humanities + Energy + \
        Veterinary + General"
model = smf.logit(formula = eq, data = reg_data.loc[list(pairs.keys()) + list(pairs.values())].sample(frac=1)).fit()

# for count DV.
# smf.glm(formula = eq, data=reg_data, family=sm.families.NegativeBinomial()).fit()

model.summary()

print(model.summary().as_latex())

eq = "self_promotion ~ C(gender, Treatment(reference='Male'))"
model = smf.logit(formula = eq, data = reg_data.loc[list(pairs.keys()) + list(pairs.values())].sample(frac=1)).fit()

model.summary()

# print(model.summary().as_latex())
