==> 5_reg_data_country_author_first_pub_year.ipynb <==
import os
import math
import json
import csv
import time
import random
import requests
import numpy as np
import pandas as pd
import seaborn as sns
import reverse_geocode
import matplotlib.colors
import matplotlib.cm as cm
import matplotlib.pyplot as plt
import matplotlib.ticker as ticker
from collections import defaultdict, Counter
from datetime import datetime

Data_Root = '/Data/Promotion/'

# pd.options.display.precision = 20

CIs = {'90': 1.645, '95': 1.96, '99': 2.576}

def yield_one_line(filename, delimiter = '\t', quote = csv.QUOTE_NONE):
    '''a generator which produce one line of a given file'''
    with open(filename, 'r') as file:
        print('processing %s...' %(filename))
        reader = csv.reader(file, delimiter=delimiter, quoting=quote)
        for row in reader:
            yield row

labels = ['Male', 'Female']

colors = sns.color_palette()[:len(labels)]

utype = {'author_id': str, 'matched_tid': str, 'matched_tid_original': str, 'matched_tid_retweet': str}
reg_data = pd.read_csv(Data_Root + "revision/reg_data_drop_missing.csv", header=0, dtype=utype)

reg_data.shape

reg_data.head(3)

Add author first publication year

aid_first_pub_year = {}

with open(Data_Root + 'revision/aid_year_paper_list.json', 'r') as ifile:
    for line in ifile:
        line = json.loads(line)
        aid = line['aid']
        first_year = sorted(line['year_pids'])[0]
        aid_first_pub_year[aid] = first_year

reg_data['aut_first_pub_year'] = reg_data['author_id'].map(aid_first_pub_year)

reg_data['aut_first_pub_year'].value_counts()

reg_data['no_years_after_first_pub'] = reg_data['pub_year'] - reg_data['aut_first_pub_year']

# 14978 rows
reg_data.loc[reg_data['no_years_after_first_pub'] < 0, 'no_years_after_first_pub'] = 0

out, bins = pd.qcut(reg_data['no_years_after_first_pub'], q=11, labels=False, duplicates='drop', retbins=True)


np.set_printoptions(suppress=True)
bins

reg_data['author_career_year_cate'] = pd.qcut(reg_data['no_years_after_first_pub'], q=11, labels=False, duplicates='drop')


reg_data.author_career_year_cate.value_counts()

Add author affiliation country

affi_rank = {}
affi_country = {}
affi_name = {}
for line in yield_one_line(Data_Root+'Affiliations.txt'):
    affi_id, rank, dname, lat, lon = line[0], line[1], line[3], line[9], line[10]
    affi_rank[affi_id] = int(rank)
    affi_name[affi_id] = dname
    if lat != "" and lon != "":
        lat, lon = float(lat), float(lon)
        res = reverse_geocode.search([(lat, lon)])
        country = res[0]['country']
        affi_country[affi_id] = country

def get_affi_country(affis):
    affi_li = affis.split('|')
    cns = []
    for affi_id in affi_li:
        # many affi ids in MAG do not have latitude and longitude info for us to infer their country
        if affi_id in affi_country:
            country = affi_country[affi_id]
            cns.append(country)
    if len(cns) > 0:
        if 'United States' in cns:
            return 'United States'
        else:
            return cns[0]
    else:
        return 'unknown'

reg_data['affiliation_country'] = reg_data['affiliation_ids'].apply(get_affi_country)

reg_data['affiliation_country'].value_counts()[:20]

len(reg_data)

reg_data.to_csv(Data_Root+"revision/reg_data_drop_missing.csv", index=False, header=True, encoding='utf-8')


utype = {'author_id': str, 'matched_tid': 'str', 'matched_tid_original': 'str', 'matched_tid_retweet': 'str'}
reg_data = pd.read_csv(Data_Root+"revision/reg_data_drop_missing.csv", header=0, dtype=utype)


len(reg_data)
